rustc_log

Trait BuildSubscriberRet

pub trait BuildSubscriberRet:
    Subscriber
    + for<'span> LookupSpan<'span>
    + Send
    + Sync { }

Trait alias for the complex return type of build_subscriber in init_logger_with_additional_layer. A Registry with any composition of tracing::Subscribers (e.g. Registry::default().with(custom_layer)) should be compatible with this type. Having an alias is also useful so rustc_driver_impl does not need to explicitly depend on tracing_subscriber.